internal/entity/user.go
package entity

import (
	"github.com/Luccas1/api-golang/pkg/entity"
	"golang.org/x/crypto/bcrypt"
)

type User struct {
	ID       entity.ID `json:"id"`
	Name     string    `json:"name"`
	Email    string    `json:"email"`
	Password string    `json:"-"` //password nunca será exibida para usuario final
}

func NewUser(name, email, password string) (*User, error) {

	// transforma o password passado em um hash
	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
	if err != nil {
		panic(err)
	}

	return &User{
		ID:       entity.NewID(),
		Name:     name,
		Email:    email,
		Password: string(hash),
	}, nil

}

// vai comparar se o hash do password salvo no user é IGUAL ao que foi passado pelo user
func (u *User) ValidatePassword(password string) bool {
	err := bcrypt.CompareHashAndPassword([]byte(u.Password), []byte(password))
	return err == nil
}
